feat[cli]: clap derive feature
