+++ Cargo.toml
@@ -2,4 +2,5 @@
 members = [
     "crates/bindings",
     "crates/clairvoyance",
+    "crates/cli",
 ]
