--- crates/clairvoyance/src/cli/cli.yml
@@ -1,22 +1,0 @@
-# Run `cargo run -- -h` for help.
-name: arbiter
-version: "0.0.1"
-args:
-  - token0:
-      long: token0
-      help: Specifies the first token for a token pair.
-      takes_value: true
-      default_value: "ETH"
-  - token1:
-      long: token1
-      help: Specifies the second token for a token pair which will be the numeraire.
-      takes_value: true
-      default_value: "USDC"
-  - fee:
-      long: fee
-      help: Specifies the basis points for the pool. [1, 5, 30, 100]
-      takes_value: true
-      default_value: "5"
-  - api_key:
-      long: api_key
-      help: Provide an Etherscan API key if function calls require it (e.g., providing a token that is not in the database).
